rust: extract Decodable trait
Extracts common functionality between the message and sequence decoders
into a `Decodable` trait that provides common `Event` handling behavior.
